+++ CHANGES.md
@@ -1,6 +1,7 @@
-## Unreleased
+## 0.2.0 (2025-11-26)
 - Update refresh command to handle users and groups via the @<user_or_group> target syntax
 - Add --version flag to passage command
+- Fix the command description for `new`
 
 ## 0.1.8 (2025-10-21)
 - deep refactor of the code